requirements.txt
Pillow==10.0.1
numpy==1.24.3
cryptography==41.0.7
tkinter-tooltip==2.0.0
email-validator==2.0.0
